tv: move activities to their own package
